Make computed properties remember the last value of this.
It stores the value of this from the last time the computed
property was evaluated.

It allows notify to re-evaluate the computed property using the
correct value for this, instead of the global object. This is
particularily necessary when a dependency notifies.

It allows a single computed property to be applied with different
values of this. This will be useful particularily to apply the
same computed property to each object in an array.